really minor change (#1)
* really minor change

Just added a | text-decoration : none | to the "a' tag. just a really minor change, you can ignore this commit if you like the way it is, I wanted to ask you about this but I just thought it would be weird. This is my first time making a commit so I'm sorry if this isn't formal. Anyway Cheers bro! I really like your videos and I hope you make a lot in the future

* Change spacing

---------

Co-authored-by: Kevin Junyang Cui <kevinjycui@protonmail.com>